refactored queries into individual functions
