+++ src/data/queries/doiQuery.ts
@@ -1,144 +1,28 @@
 import { gql } from '@apollo/client'
+import apolloClient from 'src/utils/apolloClient/apolloClient'
 import { WorkMetadata, Work } from 'src/data/types'
+import { workFragment } from 'src/data/queries/queryFragments'
 
-export const workConnection = gql`
-  fragment WorkConnectionFragment on WorkConnectionWithTotal {
-    totalCount
-    totalContentUrl
-    totalOpenLicenses
-    pageInfo {
-      endCursor
-      hasNextPage
-    }
-    resourceTypes {
-      id
-      title
-      count
-    }
-    openLicenseResourceTypes {
-      id
-      title
-      count
-    }
-    published {
-      id
-      title
-      count
-    }
-    languages {
-      id
-      title
-      count
-    }
-    fieldsOfScience {
-      id
-      title
-      count
-    }
-    licenses {
-      id
-      title
-      count
-    }
-    registrationAgencies {
-      id
-      title
-      count
-    }
-    authors {
-      id
-      title
-      count
-    }
-    creatorsAndContributors {
-      id
-      title
-      count
-    }
-    personToWorkTypesMultilevel {
-      id
-      title
-      count
-      inner {
-        id
-        title
-        count
-      }
-    }
-  }
-`
 
-export const workFragment = gql`
-  fragment WorkFragment on Work {
-    id
-    doi
-    identifiers {
-      identifier
-      identifierType
-      identifierUrl
-    }
-    types {
-      resourceTypeGeneral
-      resourceType
-    }
-    titles {
-      title
-    }
-    creators {
-      id
-      name
-      givenName
-      familyName
-      affiliation {
-        id
-        name
-      }
-    }
-    descriptions {
-      description
-      descriptionType
-    }
-    publicationYear
-    publisher {
-      name
-    }
-    version
-    container {
-      identifier
-      identifierType
-      title
-    }
-    repository {
-      id: uid
-      name
-    }
-    rights {
-      rights
-      rightsUri
-      rightsIdentifier
-      rightsIdentifierScheme
-    }
-    claims {
-      state
-    }
-    fieldsOfScience {
-      id
-      name
-    }
-    language {
-      id
-      name
-    }
-    registrationAgency {
-      id
-      name
-    }
-    registered
-    citationCount
-    viewCount
-    downloadCount
-  }
-`
+export async function fetchDoiMetadata(id: string) {
+  const { data } = await apolloClient.query<MetadataQueryData, MetadataQueryVar>({
+    query: DOI_METADATA_QUERY,
+    variables: { id },
+    errorPolicy: 'all'
+  })
+
+  return { data }
+}
+
+export async function fetchDoi(variables: QueryVar) {
+  const { data, error } = await apolloClient.query<QueryData, QueryVar>({
+    query: DOI_QUERY,
+    variables,
+    errorPolicy: 'all'
+  })
+
+  return { data, error }
+}
 
 export const DOI_METADATA_QUERY = gql`
   query getMetadataQuery($id: ID!) {
@@ -211,133 +95,6 @@ export const DOI_QUERY = gql`
       }
     }
   }
-  ${workFragment}
-`
-
-export const RELATED_CONTENT_QUERY = gql`
-  query getRelatedContentDoiQuery(
-    $id: ID!
-    $filterQuery: String
-    $cursor: String
-    $published: String
-    $resourceTypeId: String
-    $fieldOfScience: String
-    $language: String
-    $license: String
-    $registrationAgency: String
-    $repositoryId: String
-  ) {
-    work(id: $id) {
-      doi
-      types {
-        resourceTypeGeneral
-        resourceType
-      }
-      allRelated(
-        first: 25
-        query: $filterQuery
-        after: $cursor
-        published: $published
-        resourceTypeId: $resourceTypeId
-        fieldOfScience: $fieldOfScience
-        language: $language
-        license: $license
-        registrationAgency: $registrationAgency
-        repositoryId: $repositoryId
-      ) {
-        ...WorkConnectionFragment
-        nodes {
-          ...WorkFragment
-        }
-      }
-      citations(
-        first: 25
-        query: $filterQuery
-        after: $cursor
-        published: $published
-        resourceTypeId: $resourceTypeId
-        fieldOfScience: $fieldOfScience
-        language: $language
-        license: $license
-        registrationAgency: $registrationAgency
-        repositoryId: $repositoryId
-      ) {
-        ...WorkConnectionFragment
-        nodes {
-          ...WorkFragment
-        }
-      }
-      references(
-        first: 25
-        query: $filterQuery
-        after: $cursor
-        published: $published
-        resourceTypeId: $resourceTypeId
-        fieldOfScience: $fieldOfScience
-        language: $language
-        license: $license
-        registrationAgency: $registrationAgency
-        repositoryId: $repositoryId
-      ) {
-        ...WorkConnectionFragment
-        nodes {
-          ...WorkFragment
-        }
-      }
-      parts(
-        first: 25
-        query: $filterQuery
-        after: $cursor
-        published: $published
-        resourceTypeId: $resourceTypeId
-        fieldOfScience: $fieldOfScience
-        language: $language
-        license: $license
-        registrationAgency: $registrationAgency
-        repositoryId: $repositoryId
-      ) {
-        ...WorkConnectionFragment
-        nodes {
-          ...WorkFragment
-        }
-      }
-      partOf(
-        first: 25
-        query: $filterQuery
-        after: $cursor
-        published: $published
-        resourceTypeId: $resourceTypeId
-        fieldOfScience: $fieldOfScience
-        language: $language
-        license: $license
-        registrationAgency: $registrationAgency
-        repositoryId: $repositoryId
-      ) {
-        ...WorkConnectionFragment
-        nodes {
-          ...WorkFragment
-        }
-      }
-      otherRelated(
-        first: 25
-        query: $filterQuery
-        after: $cursor
-        published: $published
-        resourceTypeId: $resourceTypeId
-        fieldOfScience: $fieldOfScience
-        language: $language
-        license: $license
-        registrationAgency: $registrationAgency
-        repositoryId: $repositoryId
-      ) {
-        ...WorkConnectionFragment
-        nodes {
-          ...WorkFragment
-        }
-      }
-    }
-  }
-  ${workConnection}
   ${workFragment}
 `
