+++ src/data/queries/organizationQuery.ts
@@ -1,6 +1,27 @@
 import { gql } from '@apollo/client'
 import { Organization, OrganizationMetadata } from 'src/data/types'
-import { workConnection, workFragment } from './doiQuery'
+import apolloClient from 'src/utils/apolloClient/apolloClient'
+
+
+export async function fetchOrganizationMetadata(id: string) {
+  const { data } = await apolloClient.query<MetadataQueryData, MetadataQueryVar>({
+    query: ORGANIZATION_METADATA_QUERY,
+    variables: { id },
+    errorPolicy: 'all'
+  })
+
+  return { data }
+}
+
+export async function fetchOrganization(variables: QueryVar) {
+  const { data, error } = await apolloClient.query<QueryData, QueryVar>({
+    query: ORGANIZATION_QUERY,
+    variables,
+    errorPolicy: 'all'
+  })
+
+  return { data, error }
+}
 
 
 export const ORGANIZATION_METADATA_QUERY = gql`
@@ -61,48 +82,6 @@ export const ORGANIZATION_QUERY = gql`
     }
   }
 `
-
-export const RELATED_CONTENT_QUERY = gql`
-  query getRelatedContentQuery(
-    $id: ID
-    $gridId: ID
-    $crossrefFunderId: ID
-    $cursor: String
-    $filterQuery: String
-    $published: String
-    $resourceTypeId: String
-    $fieldOfScience: String
-    $language: String
-    $license: String
-    $registrationAgency: String
-  ) {
-    organization(
-      id: $id
-      gridId: $gridId
-      crossrefFunderId: $crossrefFunderId
-    ) {
-      works(
-        first: 25
-        after: $cursor
-        query: $filterQuery
-        published: $published
-        resourceTypeId: $resourceTypeId
-        fieldOfScience: $fieldOfScience
-        language: $language
-        license: $license
-        registrationAgency: $registrationAgency
-      ) {
-        ...WorkConnectionFragment
-        nodes {
-          ...WorkFragment
-        }
-      }
-    }
-  }
-  ${workConnection}
-  ${workFragment}
-`
-
 
 export interface MetadataQueryVar {
   id: string
