+++ src/data/queries/repositoryQuery.ts
@@ -1,6 +1,27 @@
 import { gql } from '@apollo/client'
 import { Repository, RepositoryMetadata } from 'src/data/types'
+import apolloClient from 'src/utils/apolloClient/apolloClient'
 
+
+export async function fetchRepositoryMetadata(id: string) {
+  const { data, error } = await apolloClient.query<QueryData, QueryVar>({
+    query: REPOSITORY_METADATA_QUERY,
+    variables: { id },
+    errorPolicy: 'all'
+  })
+
+  return { data, error }
+}
+
+export async function fetchRepository(variables: QueryVar) {
+  const { data, error } = await apolloClient.query<QueryData, QueryVar>({
+    query: REPOSITORY_QUERY,
+    variables,
+    errorPolicy: 'all'
+  })
+
+  return { data, error }
+}
 
 export const REPOSITORY_METADATA_QUERY = gql`
   query repositoryDetailQuery(
